Ensemble Kalman Filter

References

[1] Mandel, J. (2006). Efficient Implementation of the Ensemble Kalman Filter.

[2] M. Katzfuss, J. R. Stroud, and C. K. Wikle, “Understanding the Ensemble Kalman Filter,” The American Statistician, vol. 70, no. 4, pp. 350–357, 2016.

Methods

EnKF Prediction

PhDSE.enkf_predict — Function

enkf_predict(ensemble, Φ, process_noise_dist, [u])

Prediction step in an Ensemble Kalman filter (EnKF).

Arguments

ensemble::AbstractMatrix: the current analysis ensemble

Φ::AbstractMatrix: transition matrix, i.e. dynamics of the state space model

process_noise_dist::MvNormal: Multivariate Gaussian process-noise distribution.

u::AbstractVector (optional): affine control input to the dynamics

References

[1] Mandel, J. (2006). Efficient Implementation of the Ensemble Kalman Filter.

PhDSE.enkf_predict! — Function

enkf_predict!(c, Φ, process_noise_dist, [u])

In-place prediction step in an Ensemble Kalman filter (EnKF).

Arguments

c::FilteringCache: Cache holding pre-allocated matrices and vectors

Φ::AbstractMatrix: transition matrix, i.e. dynamics of the state space model

process_noise_dist::MvNormal: Multivariate Gaussian process-noise distribution.

u::AbstractVector (optional): affine control input to the dynamics

References

[1] Mandel, J. (2006). Efficient Implementation of the Ensemble Kalman Filter.

EnKF Correction

PhDSE.enkf_correct — Function

enkf_correct(forecast_ensemble, H, measurement_noise_dist, y, [v])

Correction step in an Ensemble Kalman filter (EnKF).

Arguments

forecast_ensemble::AbstractMatrix: the current forecast ensemble

H::AbstractMatrix: measurement matrix of the state space model

measurement_noise_dist::MvNormal: Multivariate Gaussian measurement-noise distribution.

y::AbstractVector: a measurement (data point)

v::AbstractVector (optional): affine control input to the measurement

References

[1] Mandel, J. (2006). Efficient Implementation of the Ensemble Kalman Filter. Section 2.

PhDSE.enkf_correct! — Function

enkf_correct!(c, H, measurement_noise_dist, y, [v])

In-place correction step in an Ensemble Kalman filter (EnKF).

Arguments

c::FilteringCache: Cache holding pre-allocated matrices and vectors

H::AbstractMatrix: measurement matrix of the state space model

measurement_noise_dist::MvNormal: Multivariate Gaussian measurement-noise distribution.

y::AbstractVector: a measurement (data point)

v::AbstractVector (optional): affine control input to the measurement

References

[1] Mandel, J. (2006). Efficient Implementation of the Ensemble Kalman Filter. Section 2.

Observation-matrix-free version

PhDSE.enkf_matrixfree_correct — Function

enkf_matrixfree_correct(forecast_ensemble, HX, HA, measurement_noise_dist, y; [A], [R_inverse])

Observation-matrix-free correction step in an Ensemble Kalman filter (EnKF).

Arguments

forecast_ensemble::AbstractMatrix: the current forecast ensemble

HX::AbstractMatrix: measured forecast ensemble

HA::AbstractMatrix: measured and centered forecast ensemble

measurement_noise_dist::MvNormal: Multivariate Gaussian measurement-noise distribution.

y::AbstractVector: a measurement (data point)

A::AbstractMatrix (optional): centered forecast ensemble

R_inverse::AbstractMatrix (optional): the inverse $R^{-1}$ of the measurement covariance matrix. missing per default. If R_inverse is not missing, then an alternative computation of the analysis ensemble is chosen, which uses the matrix-inversion lemma to scale as O(N^3), instead of O(d^3)!

References

[1] Mandel, J. (2006). Efficient Implementation of the Ensemble Kalman Filter. Sections 3, 4.

PhDSE.enkf_matrixfree_correct! — Function

enkf_matrixfree_correct!(c, HX, HA, A, measurement_noise_dist, y; [R_inverse])

In-place observation-matrix-free correction step in an Ensemble Kalman filter (EnKF).

Note: The arguments A, HX, HA can be computed using A_HX_HA!

Arguments

c::FilteringCache: Cache holding pre-allocated matrices and vectors

HX::AbstractMatrix: measured forecast ensemble $\mathrm{H}\mathrm{X} + v$

HA::AbstractMatrix: measured and centered forecast ensemble $\mathrm{H}\mathrm{A} + v$

A::AbstractMatrix: centered forecast ensemble $\mathrm{A} = \mathrm{X} - \bar{\mathrm{X}}$

measurement_noise_dist::MvNormal: Multivariate Gaussian measurement-noise distribution.

y::AbstractVector: a measurement (data point)

R_inverse::AbstractMatrix (optional): the inverse $R^{-1}$ of the measurement covariance matrix. missing per default. If R_inverse is not missing, then an alternative computation of the analysis ensemble is chosen, which uses the matrix-inversion lemma to scale as O(N^3), instead of O(d^3)!

References

[1] Mandel, J. (2006). Efficient Implementation of the Ensemble Kalman Filter. Sections 3, 4.

Auxiliary methods

Compute moments from an ensemble

PhDSE.ensemble_mean — Function

ensemble_mean(ens)

Compute the mean $\bar{\mathrm{X}}$ of an ensemble of states $\mathrm{X} \in \mathbb{R}^{D \times N}$

PhDSE.ensemble_mean! — Function

ensemble_mean!(out_m, ens)

Compute the mean $\bar{\mathrm{X}}$ of an ensemble of states $\mathrm{X} \in \mathbb{R}^{D \times N}$

PhDSE.centered_ensemble — Function

centered_ensemble(ens)

Compute the mean-centered ensemble $\mathrm{A} = \mathrm{X} - \bar{\mathrm{X}}$

PhDSE.centered_ensemble! — Function

centered_ensemble!(out_ens, out_m, ens)

Compute the mean-centered ensemble $\mathrm{A} = \mathrm{X} - \bar{\mathrm{X}}$

PhDSE.ensemble_cov — Function

ensemble_cov(ens)

Compute the sample covariance of an ensemble $\tilde{\mathrm{C}} = \frac{\mathrm{A}\mathrm{A}^\top}{N - 1}$

PhDSE.ensemble_mean_cov — Function

ensemble_mean_cov(ens)

Compute sample moments of an ensemble as in ensemble_mean and ensemble_cov.

PhDSE.ensemble_mean_cov! — Function

ensemble_mean_cov!(out_cov, out_ens, out_m, ens)

Compute sample moments of an ensemble. See ensemble_mean_cov.

Misc.

PhDSE.A_HX_HA! — Function

A_HX_HA!(c, H, [v])

In-place computation of measured forecast ensemble HX, as well as measured centered forecast ensemble HA and centered forecast ensemble A.

Note: not exported. Use as PhDSE.A_HX_HA!.

Arguments

c::FilteringCache: Cache holding pre-allocated matrices and vectors

H::AbstractMatrix: measurement matrix of the state space model

v::AbstractVector (optional): affine control input to the measurement