Kalman Filter

References

[1] Kalman, R. E. (1960). A new approach to linear filtering and prediction problems.

Methods

KF Prediction

PhDSE.kf_predict — Function

kf_predict(μ, Σ, Φ, Q, [u])

Prediction step in a Kalman filter.

For a numerically more stable version of the Kalman filter, have a look at the square-root Kalman filter at sqrt_kf_predict.

Arguments

μ::AbstractVector: the current filtering mean

Σ::AbstractMatrix: the current filtering covariance matrix

Φ::AbstractMatrix: transition matrix, i.e. dynamics of the state space model

Q::AbstractMatrix: transition covariance, i.e. process noise of the state space model

u::AbstractVector (optional): affine control input to the dynamics

PhDSE.kf_predict! — Function

kf_predict!(c, Φ, Q, [u])

In-place prediction step in a Kalman filter.

For a numerically more stable version of the Kalman filter, have a look at the square-root Kalman filter at sqrt_kf_predict!.

Arguments

c::FilteringCache: Cache holding pre-allocated matrices and vectors

Φ::AbstractMatrix: transition matrix, i.e. dynamics of the state space model

Q::AbstractMatrix: transition covariance, i.e. process noise of the state space model

u::AbstractVector (optional): affine control input to the dynamics

KF Correction

PhDSE.kf_correct — Function

kf_correct(μ⁻, Σ⁻, H, R, y, [v])

Correction step in a Kalman filter.

For a numerically more stable version of the Kalman filter, either have a look at the Joseph-form implementation kf_joseph_correct or at the square-root Kalman filter at sqrt_kf_correct.

Arguments

μ⁻::AbstractVector: the current predicted mean

Σ⁻::AbstractMatrix: the current predicted covariance matrix

H::AbstractMatrix: measurement matrix of the state space model

R::AbstractMatrix: measurement covariance, i.e. measurement noise of the state space model

y::AbstractVector: a measurement (data point)

v::AbstractVector (optional): affine control input to the measurements

PhDSE.kf_correct! — Function

kf_correct!(c, H, R, y, [v])

In-place correction step in a Kalman filter.

For a numerically more stable version of the Kalman filter, have a look at the square-root Kalman filter at sqrt_kf_correct!.

[2022-12-09] There is currently no in-place implementation of the Joseph-form correction.

Arguments

c::FilteringCache: Cache holding pre-allocated matrices and vectors

H::AbstractMatrix: measurement matrix of the state space model

R::AbstractMatrix: measurement covariance, i.e. measurement noise of the state space model

y::AbstractVector: a measurement (data point)

v::AbstractVector (optional): affine control input to the measurements

PhDSE.kf_joseph_correct — Function

kf_joseph_correct(μ⁻, Σ⁻, H, R, y, [v])

Joseph-form correction step in a Kalman filter.

This method is numerically more stable than the standard kf_correct and otherwise equivalent.

Arguments

μ⁻::AbstractVector: the current predicted mean

Σ⁻::AbstractMatrix: the current predicted covariance matrix

H::AbstractMatrix: measurement matrix of the state space model

R::AbstractMatrix: measurement covariance, i.e. measurement noise of the state space model

y::AbstractVector: a measurement (data point)

v::AbstractVector (optional): affine control input to the measurements