Square-root Kalman Filter

Note regarding Sqrt-Kalman filter implementation

The present implementation assumes right matrix-square roots of the involved covariance matrices. This means that a PSD covariance matrix $\mathrm{P}$ is decomposed as

\[\mathrm{P} = \mathrm{U}^\top \mathrm{U}.\]

This is in contrast to assuming a decomposition $\mathrm{P} = \mathrm{L} \mathrm{L}^\top$, as is also often the convention.

References

[1] Mohinder S Grewal and Angus P Andrews. Kalman filtering: Theory and Practice with MATLAB. John Wiley & Sons, 2014.

[2] Krämer, N., & Hennig, P. (2020). Stable implementation of probabilistic ODE solvers. arXiv:2012.10106.

Methods

Sqrt-KF Prediction

PhDSE.sqrt_kf_predict — Function

sqrt_kf_predict(μ, sqrt_Σ, Φ, sqrt_Q, [u])

Prediction step in a square-root Kalman filter. Computes the same posterior distribution as the standard Kalman filter (kf_predict) but is numerically more stable. Works entirely on right matrix-square-roots of the covariance matrices.

Arguments

μ::AbstractVector: the current filtering mean

sqrt_Σ::UpperTriangular: right matrix-square-root of the current filtering covariance matrix

Φ::AbstractMatrix: transition matrix, i.e. dynamics of the state space model

sqrt_Q::UpperTriangular: right matrix square root of transition covariance, i.e. process noise of the state space model

u::AbstractVector (optional): affine control input to the dynamics

References

[1] Krämer, N., & Hennig, P. (2020). Stable implementation of probabilistic ODE solvers.

PhDSE.sqrt_kf_predict! — Function

sqrt_kf_predict!(c, Φ, sqrt_Q, [u])

In-place prediction step in a square-root Kalman filter. Computes the same posterior distribution as the standard Kalman filter (kf_predict!) but is numerically more stable. Works entirely on right matrix-square-roots of the covariance matrices.

Arguments

c::FilteringCache: Cache holding pre-allocated matrices and vectors

Φ::AbstractMatrix: transition matrix, i.e. dynamics of the state space model

sqrt_Q::UpperTriangular: right matrix square root of transition covariance, i.e. process noise of the state space model

u::AbstractVector (optional): affine control input to the dynamics

References

[1] Krämer, N., & Hennig, P. (2020). Stable implementation of probabilistic ODE solvers.

Sqrt-KF Correction

PhDSE.sqrt_kf_correct — Function

sqrt_kf_correct(μ⁻, sqrt_Σ⁻, H, sqrt_R, y, [v])

Correction step in a square-root Kalman filter. Computes the same posterior distribution as the standard Kalman filter (kf_correct) but is numerically more stable. Works entirely on right matrix-square-roots of the covariance matrices.

Arguments

μ⁻::AbstractVector: the current predicted mean

sqrt_Σ⁻::UpperTriangular: right matrix-square-root of the current predicted covariance matrix

H::AbstractMatrix: measurement matrix of the state space model

sqrt_R::UpperTriangular: right matrix square root of measurement noise covariance of the state space model

y::AbstractVector: a measurement (data point)

v::AbstractVector (optional): affine control input to the measurement

References

[1]: Krämer, N., Bosch, N., Schmidt, J. & Hennig, P. (2022). Probabilistic ODE Solutions in Millions of Dimensions.

PhDSE.sqrt_kf_correct! — Function

sqrt_kf_correct!(c, H, sqrt_R, y, [v])

In-place correction step in a square-root Kalman filter. Computes the same posterior distribution as the standard Kalman filter (kf_correct!) but is numerically more stable. Works entirely on right matrix-square-roots of the covariance matrices.

Arguments

c::FilteringCache: Cache holding pre-allocated matrices and vectors

H::AbstractMatrix: measurement matrix of the state space model

sqrt_R::UpperTriangular: right matrix square root of measurement noise covariance of the state space model

y::AbstractVector: a measurement (data point)

v::AbstractVector (optional): affine control input to the measurement

References

[1]: Krämer, N., Bosch, N., Schmidt, J. & Hennig, P. (2022). Probabilistic ODE Solutions in Millions of Dimensions.