Ensemble Kalman Filter

References

[1] Mandel, J. (2006). Efficient Implementation of the Ensemble Kalman Filter.

[2] M. Katzfuss, J. R. Stroud, and C. K. Wikle, “Understanding the Ensemble Kalman Filter,” The American Statistician, vol. 70, no. 4, pp. 350–357, 2016.

Methods

PhDSE.enkf_predict! — Function

enkf_predict!(fcache, Φ, Q, [u])

Prediction step in an Ensemble Kalman filter (EnKF).

Arguments

fcache::EnKFCache: a cache holding memory-heavy objects

Φ::AbstractMatrix: transition matrix, i.e. dynamics of the state space model

u::AbstractVector (optional): affine control input to the dynamics

References

[1] Mandel, J. (2006). Efficient Implementation of the Ensemble Kalman Filter.

PhDSE.enkf_correct! — Function

enkf_correct!(fcache, H, R_inv, y, [v])

Correction step in an Ensemble Kalman filter (EnKF).

Note:

Calls omf_enkf_correct! intrinsically.

Arguments

fcache::EnKFCache: a cache holding memory-heavy objects

H::AbstractMatrix: measurement matrix of the state space model

R_inv::AbstractMatrix: inverse of the measurement noise covariance of the state space model

y::AbstractVector: a measurement (data point)

v::AbstractVector (optional): affine control input to the measurement

References

[1] Mandel, J. (2006). Efficient Implementation of the Ensemble Kalman Filter.

PhDSE.omf_enkf_correct! — Function

omf_enkf_correct!(fcache, R_inv, y)

Correction step in an observation-matrix-free (omf) Ensemble Kalman filter (EnKF), assuming that the observation matrix never has to be built.

Instead, the products $HX$ and $HA = H\cdot\left(X^f - \mathbb{E}\left[X^f\right]\right)$ are pre-computed outside of the correction function and $HX$ and $HA$ are passed to the correction function. Assuming R_inv is a Diagonal matrix and $HX$ and $HA$ are cheap to compute, this results in a correction cost that is linear in the {state,observation}-dimension.

Note:

$HX$ and $HA$ have to be stored in the fcache before calling this function!

Arguments

fcache::EnKFCache: a cache holding memory-heavy objects

R_inv::AbstractMatrix: inverse of the measurement noise covariance of the state space model

y::AbstractVector: a measurement (data point)

References

[1] Mandel, J. (2006). Efficient Implementation of the Ensemble Kalman Filter.